Refactor token api route
